Add '-e' flag to upload flags.
From the avrdude docs: "Causes a chip erase to be executed.  This will reset the contents of the flash ROM and EEPROM to the value `0xff', and clear all lock bits. Except for ATxmega devices which can use page erase, it is basically a
prerequisite command before the flash ROM can be reprogrammed again. The only exception would be if the new contents would exclusively cause bits to be programmed from the value `1' to `0'."