Work on hang error. Turns out it might all just be because of the timing being off because of the serial output!?
The response queueing I tried to implement probably violates the USB spec. Section 5.5.5 says:

If a Setup transaction is received by an endpoint before a previously initiated control transfer is completed, the device must abort the current transfer/operation and handle the new control Setup transaction. A Setup transaction should not normally be sent before the completion of a previous control transfer. However, if a transfer is aborted, for example, due to errors on the bus, the host can send the next Setup transaction prematurely from the endpoint’s perspective.